refactor: create packages for cucumber and pact
